Extracted TidCarouselV2 and CurateV2 components into own files.
